pyproject.toml
[tool.poetry]
name = "fullctl"
repository = "https://github.com/fullctl/fullctl"
version = "1.15.0-rc.75"
description = "Core classes and functions for service applications"
authors = ["FullCtl <code@fullctl.com>"]
readme = "README.md"
license = "Apache-2.0"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Topic :: Software Development",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

packages = [
    { include = "fullctl", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.9"

# django deps
django = ">=3.2"
# >3.5.1 is not backwards compatible
django-autocomplete-light = ">=3.9.7"
# needs to allow 2.0 until django 4.2 fixes are complete
django-crispy-forms = ">=2.0"
django-grainy = ">=2"
django-handleref = ">=2"
django-inet = ">=1"
django-recaptcha = ">=2"
djangorestframework = ">=3"
django-reversion = ">=5"
django-structlog = ">=2.1.3"
# TODO: django-filter 25.1 no longer supports DRF's autoschema generation
# so either it or the schema generator needs to be replaced (2025-04-17)
django-filter = ">=23.4,<25"
django-cors-headers = ">=4.3.1"

pydantic = ">=2.6.3"

# can't release to pypi with git reference
# social-auth-app-django = { git = "https://github.com/python-social-auth/social-app-django.git", branch="master" }

social-auth-app-django = ">=5.2"

setuptools = "*"

django-netfields = "<2"
# netaddr: django-netfields already installs this, but we are now
# also using it directly to validate mac addresses
netaddr = ">=1.2.1"

# prod serving
whitenoise = ">=6"
pyuwsgi = ">=2.0.19"

confu = ">=1.9"
grainy = ">=1.8.1"
psycopg = {extras = ["binary"], version = ">=3.1.8"}
# force update
pyyaml = ">=5"
munge = ">=1.3.0"
django-countries = ">=7.6"

# for drf swagger
inflection = ">=0.5.1"

[tool.poetry.dev-dependencies]
# tests
coverage = ">=5"
pytest = ">=6"
pytest-django = ">=3.8"
pytest-cov = "*"
requests-mock = ">=1.11"
tox = ">=3.20"
tox-gh-actions = ">=2.9.1"

# linting
black = ">=20"
isort = ">=5.7"
flake8 = ">=3.8"
pre-commit = ">=2.13"
pyupgrade = ">=2.19"

# docs
markdown = "*"
markdown-include = ">=0.5"
mkdocs = ">=1.2.3"
uritemplate = "*"

# ctl
ctl = { git = "https://github.com/20c/ctl.git", branch = "main" }
jinja2 = ">=3"
tmpl = ">=1"

[build-system]
requires = ["poetry-core>=1"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."markdown.extensions"]
pymdgen = "pymdgen.md:Extension"

[tool.isort]
profile = "black"
multi_line_output = 3
